RL im GUI erweitert. Weiteres Trainieren jetzt auch über GUI möglich. Bugfix in envoirment.py

+++ Reinforcement_Learning/enviroment.py
@@ -50,14 +50,14 @@ class enviroment(gym.Env):
             low=np.array([0, 0, 0, 0, 0, 0]),
             high=np.array([enviroment_cfg["speed_scaling"], 300, 300, 300, 300, 300]),
             shape=(6,),
-            dtype=np.float64
+            dtype=np.float32
         )
 
         # Action definieren
         self.action_space = spaces.Box(
             low=np.array([-1.0, -1.0]),
             high=np.array([1.0, 1.0]),
-            dtype=np.float64
+            dtype=np.float32
         )
 
     # State abfragen
